fix(core): Patch circular import + partial gene download in precompute
Due to the circular import, and common functionality, the lineage and mutation packages were grouped.

BREAKING CHANGE: Breaks mutation functionality for multiple samples
