--- alvoc/core/lineages/__init__.py
@@ -1,1 +1,0 @@
-from alvoc.core.lineages.analyze import find_lineages
