+++ alvoc/core/variants/lineages/__init__.py
@@ -1,0 +1,1 @@
+from alvoc.core.variants.lineages.main import find_lineages
