+++ alvoc/core/variants/mutations/__init__.py
@@ -1,0 +1,1 @@
+from alvoc.core.variants.mutations.main import find_mutants, find_mutants_in_bam
